backend/requirements.txt
requests==2.31.0
flask==2.3.3
flask-cors==4.0.0
textblob==0.17.1
praw==7.7.1
beautifulsoup4==4.12.2
newsapi-python==0.2.7
mysql-connector-python==8.2.0
lxml==4.9.3